Modularized object detection
